% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{add_msoa_names}
\alias{add_msoa_names}
\title{If our \code{lookup} is MSOA, add the House of Commons Library MSOA Names -}
\usage{
add_msoa_names(x)
}
\description{
If our \code{lookup} is MSOA, add the House of Commons Library MSOA Names -
}
\keyword{internal}
