% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{api_data_req}
\alias{api_data_req}
\title{Build an API request for table data or spatial data}
\usage{
api_data_req(ids, url, fields, geo = TRUE, crs = NULL, ...)
}
\arguments{
\item{ids}{The IDs of the data to be requested}

\item{url}{The API URL to query}

\item{fields}{Which fields to include in the table returned}

\item{geo}{Whether this query should "returnGeometry" (TRUE) or not}

\item{crs}{The Coordinate Reference System (CRS) code to use}

\item{...}{Any arguments to be passed to \code{opengeo_api_req()}}
}
\description{
Build an API request for table data or spatial data
}
\keyword{internal}
