% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geometry_to_table.R
\name{arrange_service_names_by_res_codes}
\alias{arrange_service_names_by_res_codes}
\title{Arrange geo schema by preferred resolutions (only applies when no user pref)}
\usage{
arrange_service_names_by_res_codes(x, r = res_codes())
}
\arguments{
\item{x}{schema tibble}

\item{r}{vector of resolution codes e.g. "BGC" "BFE"}
}
\description{
Arrange geo schema by preferred resolutions (only applies when no user pref)
}
