% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundr-package.R
\name{base_url}
\alias{base_url}
\title{The base part of the OpenGeography API standard URLs, for convenience}
\usage{
base_url()
}
\description{
The base part of the OpenGeography API standard URLs, for convenience
}
\keyword{internal}
