% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundr-package.R
\name{batch_it}
\alias{batch_it}
\title{Batch a vector or list into a list of elements with a maximum size}
\usage{
batch_it(x, batch_size)
}
\arguments{
\item{x}{A vector or list}

\item{batch_size}{integer. The size (length) of batches to create}
}
\value{
A list
}
\description{
Batch a vector or list into a list of elements with a maximum size
}
\examples{
\dontrun{batch_it(letters, 6L)}
}
\keyword{internal}
