% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundr-package.R
\docType{package}
\name{boundr-package}
\alias{boundr}
\alias{boundr-package}
\title{boundr: Retrieve Area Boundaries and Data from the ONS Open Geography Portal}
\description{
The main purpose of this package is to download area lookups and boundaries (in GeoJSON format) using the ONS Open Geography API, for all areas at a specified level within a specified area.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/francisbarton/boundr}
  \item Report bugs at \url{https://github.com/francisbarton/boundr/issues}
}

}
\author{
\strong{Maintainer}: Francis Barton \email{fbarton@alwaysdata.net}

}
\keyword{internal}
