% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{boundr_options}
\alias{boundr_options}
\alias{opts}
\title{Use this to set custom options for \code{bounds()}}
\usage{
boundr_options(
  resolution = res_codes(),
  return_width = c("tidy", "full", "minimal"),
  crs = 4326,
  query_option = NULL
)

opts(
  resolution = res_codes(),
  return_width = c("tidy", "full", "minimal"),
  crs = 4326,
  query_option = NULL
)
}
\arguments{
\item{resolution}{character. See \code{res_codes()} for the potential options. How
generalised should the boundary be, and whether coastal boundaries should
adhere to the coastline or to the full territorial extent. BGC by default
(G = Generalised (20m), C = limited to the coastline.) F indicates Full
resolution; S indicates Super-generalised (200m); U indicates Ultra-
generalised (500m) boundary resolution. Use E instead of C for full
extent boundaries (e.g. BFE). Not all combinations are available.
Ignored if \code{geometry} is set to "centroids".}

\item{return_width}{character. How many of the possible columns in the
returned table to keep. Options are "tidy", "full" or "minimal". "Tidy" aims
to return four data columns (usually) - two columns for the lookup level
codes and names, and two for the within level codes and names. Plus a
geometry column. "Full" aims to return all data columns from the lookup.
"Minimal" aims to return just the two (usually!) data columns relating to
\code{lookup_level}.
If \code{within} is not supplied then "tidy" will be equivalent to "minimal".}

\item{crs}{The Coordinate Reference System (CRS) code to use}

\item{query_option}{numeric. Defaults to 1, which means that the URL will
just be the first one from the list of possible services resulting from the
level and year filters above. If this does not give you what you want, you
can run the function again with a different option from the list.}
}
\description{
\code{opts()} is an alias for this function.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
boundr_options(crs = 27700) # Set the CRS to British National Grid
boundr_options(return_width = "full") # Ask boundr to return all data columns
\dontshow{\}) # examplesIf}
}
