% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_api_query.R
\name{build_api_query}
\alias{build_api_query}
\title{Helper Function to Build Queries to the Open Geography API}
\usage{
build_api_query(ref, where_level, where = NULL, fields = "*", sr = 4326)
}
\arguments{
\item{ref}{an integer, passed by the calling function, that
indicates which url_string to use.}

\item{where_level}{The variable name associated with the
locations filter. e.g. \code{"lsoa11nm"} or \code{"rgn20cd"}}

\item{where}{A vector of names or codes to filter the data by. If nothing is
provided then the full table should be returned.}

\item{fields}{The fields of the data to be returned. Defaults to \code{"*"}
(all); can instead be a set of column names/variables.}

\item{sr}{The (EPSG) spatial reference of any returned geometry.
4326 ("WGS 84") by default. Can be specified as numeric or character.}
}
\value{
a string that should function as a valid API query
}
\description{
A function called by create_custom_lookup.R and geo_get.R to build a
valid query.
}
\examples{
build_api_query(ref = 4,
  where_level = "cauth20nm",
  where = "Greater Manchester",
  fields = c("lad20cd", "lad20nm", "cauth20cd", "cauth20nm")
)
build_api_query(ref = 12,
  where_level = "lad20nm",
  where = c(
    "Cheltenham", "Gloucester",
    "Stroud", "Cotswold",
    "Tewkesbury", "Forest of Dean"
  ),
  fields = c("lad20cd", "lad20nm")
)
}
