% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fields.R
\name{build_fields}
\alias{build_fields}
\title{filter a lookup table to get formal field codes from informal inputs}
\usage{
build_fields(x)
}
\arguments{
\item{x}{a string or vector of strings}
}
\description{
filter a lookup table to get formal field codes from informal inputs
}
\keyword{internal}
