% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_query_info.R
\name{build_flat_query}
\alias{build_flat_query}
\title{This is such a crucial function to the whole package! But so simple.}
\usage{
build_flat_query(var, vec)
}
\description{
This is such a crucial function to the whole package! But so simple.
}
\keyword{internal}
