% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_lookup.R
\name{common_lookup}
\alias{common_lookup}
\title{Common procedure for joinable and not-necessarily-joinable lookup()s}
\usage{
common_lookup(
  lookup_level,
  within_level = NULL,
  within_names = NULL,
  within_codes = NULL,
  lookup_year = NULL,
  within_year = NULL,
  opts = boundr_options(),
  joinable = FALSE
)
}
\arguments{
\item{joinable}{logical. Whether this query is joinable (\code{TRUE}) or
forms part of a geospatial query (\code{FALSE}). In the latter case, this
function will try to return a lookup table that contains the appropriate
field for spatial data (that is, boundaries or centroids) to be joined onto.
For \code{{boundr}}'s internal use only, users should ignore this.}
}
\description{
Common procedure for joinable and not-necessarily-joinable lookup()s
}
\keyword{internal}
