% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_custom_lookup.R
\name{create_custom_lookup}
\alias{create_custom_lookup}
\title{Uses build_api_query() to Retrieve, Tidy and Return a Lookup Table}
\usage{
create_custom_lookup(
  bounds_level,
  within,
  within_level,
  include_msoa = NULL,
  return_style = "tidy",
  include_welsh_names = NULL
)
}
\arguments{
\item{bounds_level}{The lowest level at which to return codes and names, eg
"LSOA". Has to be one of "lsoa", "msoa", "wd/ward", "lad",
"cty/county". Case-insensitive.}

\item{within}{The name of a geographic area to filter by eg "Swindon",
"Gloucestershire", "Wales".}

\item{within_level}{Upper geographic level to filter at. eg if filtering to
find all LSOAs in a local authority, \code{within_level} will be "lad". Has
to be one of "wd/ward", "lad", "cty/county", "utla/upper", "rgn/region",
"cauth" or "ctry/country". Case-insensitive. Not all combinations of
\code{bounds_level} and \code{within_level} make sense or are possible! NB
"county" includes metropolitan counties such as "Inner London", "Tyne and
Wear" and "West Midlands".}

\item{include_msoa}{If \code{bounds_level} is LSOA and return_style is "tidy",
whether to also include MSOA columns (in "tidy" return style). If
\code{bounds_level} is MSOA, this will be forced to \code{TRUE}.}

\item{return_style}{"tidy" (the default) means all available columns between
\code{bounds_level} and \code{within_level} will be returned, but with any
empty columns removed. "simple" means that only the code and name (cd and
nm) columns for \code{bounds_level} and \code{within_level} are returned -
other columns are omitted. "minimal" means 'only return the columns for
\code{bounds_level}'.}

\item{include_welsh_names}{Only makes a difference when \code{bounds_level} =
msoa, or when \code{bounds_level} = lsoa and \code{return_style} = "tidy".
\code{FALSE} returns no Welsh language columns. \code{TRUE} attempts to
return Welsh language LAD and MSOA names where relevant. \code{NULL} (the
default) means that an educated decision will be made by the program,
based on whether any of the areas returned have "^W" codes.}
}
\value{
a data frame (tibble)
}
\description{
Roughly equivalent to \code{geo_get(boundaries = FALSE)}
}
\examples{
\dontrun{
create_custom_lookup(
  bounds_level = "msoa",
  within = "Swindon",
  within_level = "lad",
  return_style = "simple"
)
}
\dontrun{
create_custom_lookup(
  bounds_level = "msoa",
  within = "Swansea",
  within_level = "lad",
  return_style = "tidy"
)
}
# TODO add in more (and more interesting) examples?
}
