% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lookup_table.R
\name{create_lookup_table}
\alias{create_lookup_table}
\title{Create a lookup table by querying the ONS OpenGeography API}
\usage{
create_lookup_table(
  lookup,
  within = NULL,
  within_names = NULL,
  within_codes = NULL,
  return_width = c("tidy", "full", "minimal"),
  lookup_year = NULL,
  within_year = NULL,
  country_filter = c("UK|GB|EW|EN|SC|WA", "UK", "GB", "EW", "EN", "SC", "WA"),
  option = NULL,
  standalone = TRUE,
  chatty = rlang::is_interactive()
)
}
\arguments{
\item{lookup}{character. Lower level area code eg "lsoa", "wd", "lad".
Equivalent to the \code{lookup} parameter in \code{bounds()}.}

\item{within}{character. Higher level area code eg "lad", "cty". Equivalent
to the \code{within} parameter in \code{bounds()}.}

\item{within_names, within_codes}{character. In order to restrict data
returned to a specific area, either \code{within_names} or \code{within_codes} must
be provided. Otherwise all boundaries within the country at that level will
be retrieved. Use place names eg "Essex" to restrict to a certain
geographical area. Use ONS area codes eg "W02000103" likewise (this is
useful with wards, where there are many that share identical names). A
vector of multiple names or codes can be supplied.}

\item{return_width}{character. How many of the possible columns in the
returned table to keep. Options are "tidy", "full" or "minimal".}

\item{lookup_year}{numeric or character. A specific year for data relating
to parameter \code{x}, if needed. Defaults to \code{NULL}, which will return the most
recent data.}

\item{within_year}{numeric or character. A specific year for data relating
to parameter \code{y}, if needed. Defaults to \code{NULL}, which will return the most
recent data.}

\item{country_filter}{character. Open Geography datasets are sometimes
available just within certain countries. Specify a country code if you want
your results restricted to a certain country only - eg "WA" for Wales, "EW"
for England and Wales. By default returns all options.}

\item{option}{numeric. Defaults to 1, which means that the URL will just be
the first one from the list of possible services resulting from the level
and year filters above. If this does not give you what you want, you can
run the script again with a different option from the list.}

\item{standalone}{logical. Whether this query is standalone (\code{TRUE}) or
forms part of a geospatial query (\code{FALSE}). In the latter case, this
function will try to return a lookup table that contains the appropriate
columns for spatial data (boundaries or centroids) to be joined onto.}

\item{chatty}{Boolean. Whether to print feedback on the 'decisions' the
function has taken about which table to query. Default \code{TRUE} when the
function is run in an interactive session, \code{FALSE} otherwise.}
}
\value{
A tibble
}
\description{
Calls \code{return_lookup_query_info()} and processes the output, returning a
tibble.
}
\examples{
create_lookup_table("msoa", "lad", "Swindon")
create_lookup_table("wd", "sener", lookup_year = 2022, return_width = "full")

}
