% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lookup.R
\name{extract_lookup}
\alias{extract_lookup}
\title{Helper function}
\usage{
extract_lookup(x)
}
\arguments{
\item{x}{a JSON return from an API query}
}
\value{
a data frame
}
\description{
Helper function
}
