% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_get_bounds.R
\name{geo_get_bounds}
\alias{geo_get_bounds}
\title{Return boundary geometries for supplied list of area codes}
\usage{
geo_get_bounds(
  bounds_query_level,
  area_codes,
  return_boundaries = TRUE,
  return_centroids = FALSE,
  centroid_fields = FALSE,
  shape_fields = FALSE,
  spatial_ref = 4326,
  quiet_read = TRUE
)
}
\arguments{
\item{bounds_query_level}{area level to query for, e.g. "lsoa11cd". Needs to
be the full code not an abbreviation such as "lsoa"}

\item{area_codes}{a vector of codes that match the level of
bounds_query_level. Probably to be supplied from a column pulled from a
lookup df or similar.}

\item{return_boundaries}{whether to retrieve object boundaries data from
the API. Default \code{TRUE}. If \code{return_boundaries} and
\code{return_centroids} and \code{centroid_fields} are all \code{FALSE}, a
plain summary df without any geometry will be returned.}

\item{return_centroids}{whether to retrieve area centroids instead of
boundaries. Default \code{FALSE}. If set to TRUE then it will override
\code{return_boundaries} whether that was set TRUE or otherwise. If
\code{return_boundaries} and \code{return_centroids} are both \code{FALSE},
a plain summary data frame without geometry will be returned.}

\item{centroid_fields}{Boolean, default FALSE. Whether to include BNG
eastings, northings, lat and long fields in the return. NB this \emph{doesn't}
apply to direct (population-weighted) centroid queries.}

\item{shape_fields}{Boolean, default FALSE. Whether to include
Shape__Area and Shape__Length fields in the return when returning
boundaries.}

\item{spatial_ref}{The (EPSG) spatial reference of any returned geometry.
Default value: 4326 ("WGS 84"). This parameter is ignored peacefully if
no geometry is returned/returnable, eg lookup queries}

\item{quiet_read}{Controls quiet parameter to sf::st_read}
}
\value{
an sf object
}
\description{
Return boundary geometries for supplied list of area codes
}
