% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_serious.R
\name{get_serious}
\alias{get_serious}
\title{filter a lookup table to get formal field codes from informal inputs}
\usage{
get_serious(x)
}
\arguments{
\item{x}{an input string or vector of strings}
}
\description{
filter a lookup table to get formal field codes from informal inputs
}
