% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundr-package.R
\name{gregg}
\alias{gregg}
\title{grepl a glued regex}
\usage{
gregg(x, rx, g = parent.frame())
}
\arguments{
\item{x}{A character vector to check}

\item{rx}{A string that after processing by \code{\link[glue:glue]{glue::glue_data()}} will be used
as a regex pattern in \code{\link[=grepl]{grepl()}}}
}
\value{
A logical value
}
\description{
Facilitates using regex in search/filter patterns, and puts the
arguments "the right way round" (x first, then pattern), unlike \code{\link[=grepl]{grepl()}}
}
\keyword{internal}
