% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hocl_msoa_names.R
\docType{data}
\name{hocl_msoa_names}
\alias{hocl_msoa_names}
\title{A data frame of all MSOAs in England and Wales: codes and names, with House of Commons Library MSOA names}
\format{
A data frame with 7201 rows and 5 variables:
\describe{
\item{msoa11cd}{MSOA code}
\item{msoa11nm}{MSOA name (English)}
\item{msoa11nmw}{MSOA name (Welsh)}
\item{msoa11hclnm}{MSOA HCL name (English)}
\item{msoa11hclnmw}{MSOA HCL name (Welsh)}
}
}
\usage{
hocl_msoa_names
}
\description{
A data frame of all MSOAs in England and Wales: codes and names, with House of Commons Library MSOA names
}
\keyword{datasets}
