% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundr-package.R
\name{ifnull}
\alias{ifnull}
\title{This is the same as \verb{\%||\%} but I prefer to use bracketed fns than infixes}
\usage{
ifnull(x, y)
}
\arguments{
\item{x, y}{R objects}
}
\description{
This is the same as \verb{\%||\%} but I prefer to use bracketed fns than infixes
}
\keyword{internal}
