% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jogger-package.R
\docType{package}
\name{jogger-package}
\alias{jogger}
\alias{jogger-package}
\title{jogger: Retrieve Area Boundaries and Data from the ONS Open Geography Portal}
\description{
The main function of this package is to download area boundaries (in GeoJSON format) using the ONS Open Geography API, for all sub-areas - at a specified level - within a specified area. The main script will return a data frame with the sub-area geometry as an `sf` object ready to be visualised as a map. Initially just for areas within England and Wales, with a hope to add Scottish geographies at a later date.
}
\author{
\strong{Maintainer}: Francis Barton \email{fbarton@alwaysdata.net}

}
\keyword{internal}
