% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lad21nmw_lookup.R
\docType{data}
\name{lad21nmw_lookup}
\alias{lad21nmw_lookup}
\title{A list of all LADs in England and Wales: codes and names}
\format{
A data frame with 374 rows and 3 variables:
\describe{
\item{lad21cd}{LAD code}
\item{lad21nm}{LAD name (English)}
\item{lad21nmw}{LAD name (Welsh)}
}
}
\usage{
lad21nmw_lookup
}
\description{
A list of all LADs in England and Wales: codes and names
}
\keyword{datasets}
