% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsoa11cdnm.R
\docType{data}
\name{lsoa11cdnm}
\alias{lsoa11cdnm}
\title{A list of all LSOAs in England and Wales: codes and names}
\format{
A data frame with 34753 rows and 2 variables:
\describe{
\item{lsoa11cd}{LSOA code}
\item{lsoa11nm}{LSOA name}
}
}
\usage{
lsoa11cdnm
}
\description{
A list of all LSOAs in England and Wales: codes and names
}
\keyword{datasets}
