% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsoa_to_msoa_lookup.R
\name{lsoa_to_msoa_lookup}
\alias{lsoa_to_msoa_lookup}
\title{Helper Function to Convert LSOA names to MSOA Codes and Names Including HOCL Names}
\usage{
lsoa_to_msoa_lookup(df, hocl_names_version = "1.8", nmw = TRUE, keep = FALSE)
}
\arguments{
\item{df}{data frame with at least column \code{lsoa11nm} and possibly \code{lsoa11cd}}

\item{hocl_names_version}{latest version of the House of Commons Library MSOA Names. Needs to be a string e.g. "1.8" (default)}

\item{nmw}{whether to keep the Welsh language MSOA names (ONS and HOCL) or not. Boolean. Default TRUE.}

\item{keep}{whether to keep the LSOA column(s) or just return a df starting at MSOA level. Default FALSE.}
}
\value{
a data frame (a lookup table)
}
\description{
Helper Function to Convert LSOA names to MSOA Codes and Names Including HOCL Names
}
