% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_lad21_lookup.R
\docType{data}
\name{oa_lad21_lookup}
\alias{oa_lad21_lookup}
\title{A lookup of OAs in England and Wales to higher geographies}
\format{
A data frame with 181408 rows and 15 variables:
\describe{
\item{oa11cd}{OA code}
\item{lsoa11cd}{LSOA code}
\item{lsoa11nm}{LSOA name}
\item{msoa11cd}{MSOA code}
\item{msoa11nm}{MSOA name}
\item{lad20cd}{LAD code}
\item{lad20nm}{LAD name}
\item{ltla21cd}{LTLA code}
\item{ltla21nm}{LTLA name}
\item{utla21cd}{UTLA code}
\item{utla21nm}{UTLA name}
\item{rgn21cd}{Region code}
\item{rgn21nm}{Region name}
\item{ctry21cd}{Country code}
\item{ctry21nm}{Country name}
}
}
\usage{
oa_lad21_lookup
}
\description{
A lookup of OAs in England and Wales to higher geographies
}
\keyword{datasets}
