% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundr-package.R
\name{og_}
\alias{og_}
\title{The base part of the OpenGeography API standard URLs, for convenience}
\usage{
og_()
}
\description{
The base part of the OpenGeography API standard URLs, for convenience
}
\keyword{internal}
