% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundr-package.R
\name{ogu}
\alias{ogu}
\title{The base part of the OpenGeography API standard URLs, for convenience}
\usage{
ogu()
}
\description{
The base part of the OpenGeography API standard URLs, for convenience
}
\keyword{internal}
