% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{opengeo_api_req}
\alias{opengeo_api_req}
\title{Basic OpenGeography API request}
\usage{
opengeo_api_req(
  url,
  append = "0/query",
  format = "pjson",
  user_agent = "boundr R package (https://codeberg.org/francisbarton/boundr)"
)
}
\arguments{
\item{url}{The API URL to query}

\item{append}{string to be appended to the URL path}

\item{format}{data format to return from the API, \code{"pjson"} by default}

\item{user_agent}{Allows user to specify a certain User Agent string in the
header of each query. Sends the string \code{"boundr R package"} by default}
}
\description{
Basic OpenGeography API request
}
