% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{query_opengeo_api}
\alias{query_opengeo_api}
\title{Perform an API query using a request object}
\usage{
query_opengeo_api(req)
}
\arguments{
\item{req}{An \code{httr2} request object}
}
\description{
Perform an API query using a request object
}
