% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_lookup.R
\name{remove_nmw}
\alias{remove_nmw}
\title{If we don't have any areas in Wales, remove 'nmw' columns for simplicity}
\usage{
remove_nmw(tbl)
}
\description{
If we don't have any areas in Wales, remove 'nmw' columns for simplicity
}
\keyword{internal}
