% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_spatial.R
\name{res_codes}
\alias{res_codes}
\title{A list of all available resolutions for boundary geometries in the current
OpenGeography schema. Not all resolutions are available for all area types!
The most common ones are listed first, with the "generalised" (20m
resolution) BGC being the default option if you don't specify one.}
\usage{
res_codes()
}
\description{
A list of all available resolutions for boundary geometries in the current
OpenGeography schema. Not all resolutions are available for all area types!
The most common ones are listed first, with the "generalised" (20m
resolution) BGC being the default option if you don't specify one.
}
