% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_spatial.R
\name{res_codes_regex}
\alias{res_codes_regex}
\title{Convert all res_codes() into a single bracketed regex separated by \code{|}}
\usage{
res_codes_regex()
}
\description{
Convert all res_codes() into a single bracketed regex separated by \code{|}
}
\keyword{internal}
