% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{return_bounds_data}
\alias{return_bounds_data}
\title{Perform an API query and handle the returned spatial data}
\usage{
return_bounds_data(ids, url, crs)
}
\arguments{
\item{ids}{the IDs of the data to be requested}

\item{url}{The API URL to query}

\item{crs}{the Coordinate Reference System (CRS) code to use}
}
\description{
Perform an API query and handle the returned spatial data
}
