% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_query_info.R
\name{return_lookup_table_info}
\alias{return_lookup_table_info}
\title{Get initial data about query}
\usage{
return_lookup_table_info(
  lookup,
  within_level,
  lookup_year,
  within_year,
  joinable
)
}
\description{
Get initial data about query
}
\keyword{internal}
