% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_query_info.R
\name{return_narrow_table_info}
\alias{return_narrow_table_info}
\title{Get initial data about query for "narrow" tables (with no lookup table)}
\usage{
return_narrow_table_info(lookup, lookup_year, rs = NULL)
}
\description{
Get initial data about query for "narrow" tables (with no lookup table)
}
\keyword{internal}
