% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_query_info.R
\name{return_query_ids}
\alias{return_query_ids}
\title{This function also actually queries the API}
\usage{
return_query_ids(query_url, where_string)
}
\description{
This function also actually queries the API
}
\keyword{internal}
