% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{return_result_ids}
\alias{return_result_ids}
\title{Perform an API query and extract the returned IDs
For large queries, just return IDs that can then be batched for full queries}
\usage{
return_result_ids(url, where)
}
\arguments{
\item{url}{The API URL to query}

\item{where}{any geographical filters to be applied}
}
\description{
Perform an API query and extract the returned IDs
For large queries, just return IDs that can then be batched for full queries
}
