% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{return_spatial_data}
\alias{return_spatial_data}
\title{Perform an API query and handle the returned spatial data}
\usage{
return_spatial_data(ids, url, crs = 4326, max_tries = 3, verbosity = 0, ...)
}
\arguments{
\item{ids}{the IDs of the data to be requested}

\item{url}{The API URL to query}

\item{crs}{the Coordinate Reference System (CRS) code to use. 4326 by
default.}

\item{max_tries}{integer. 3 by default (for now). Passed to \code{httr2::req_retry ()} and controls how many times to try to perform the request, should
initial attempt(s) fail.}

\item{verbosity}{integer. 0 by default. Passed to \code{httr2::req_perform()} and
controls the verbosity of the printed output from performing the request.
Can be any integer from 0 to 3, or \code{NULL}. See \code{?req_perform} for more
detail.}

\item{...}{any arguments to be passed to \code{opengeo_api_req()}}
}
\description{
Perform an API query and handle the returned spatial data
}
\keyword{internal}
