% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{return_table_data}
\alias{return_table_data}
\title{Perform an API query and handle the returned table data}
\usage{
return_table_data(ids, url, fields)
}
\arguments{
\item{ids}{the IDs of the data to be requested}

\item{url}{The API URL to query}

\item{fields}{which fields to include in the table returned}
}
\description{
Perform an API query and handle the returned table data
}
