% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{safely_query_opengeo_api}
\alias{safely_query_opengeo_api}
\title{In order to use purrr::safely() in the package. See R/zzz.R also.}
\usage{
safely_query_opengeo_api(...)
}
\description{
In order to use purrr::safely() in the package. See R/zzz.R also.
}
