% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{spatial_data_req}
\alias{spatial_data_req}
\title{Build an API request for spatial (geojson) data}
\usage{
spatial_data_req(ids, url, crs = 4326, ...)
}
\arguments{
\item{ids}{the IDs of the data to be requested}

\item{url}{The API URL to query}

\item{crs}{the Coordinate Reference System (CRS) code to use. 4326 by
default.}

\item{...}{any arguments to be passed to \code{opengeo_api_req()}}
}
\description{
Build an API request for spatial (geojson) data
}
