% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_tier_region_ctry_lookup.R
\docType{data}
\name{upper_tier_region_ctry_lookup}
\alias{upper_tier_region_ctry_lookup}
\title{A lookup of UTLAs (EW) to their Region (E) and Country}
\format{
A data frame with 331 rows and 8 variables:
\describe{
\item{ltla21cd}{LTLA code}
\item{ltla21nm}{LTLA name}
\item{utla21cd}{UTLA code}
\item{utla21nm}{UTLA name}
\item{rgn21cd}{Region code}
\item{rgn21nm}{Region name}
\item{ctry21cd}{Country code}
\item{ctry21nm}{Country name}
}
}
\usage{
upper_tier_region_ctry_lookup
}
\description{
A lookup of UTLAs (EW) to their Region (E) and Country
}
\keyword{datasets}
