% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundr-package.R
\name{batch_it}
\alias{batch_it}
\title{Batch a vector into a list}
\usage{
batch_it(x, batch_size)
}
\arguments{
\item{x}{a vector}

\item{batch_size}{the number of items in each batch of the returned list}
}
\description{
Batch a vector into a list
}
\keyword{internal}
