% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_api_query.R
\name{build_api_query}
\alias{build_api_query}
\title{Helper Function to Build Queries to the Open Geography API}
\usage{
build_api_query(
  ref,
  type = "census",
  within_level,
  within = NULL,
  fields = "*",
  return_geometry = TRUE,
  sr = 4326
)
}
\arguments{
\item{ref}{an integer, passed by the calling function, that
indicates which url_string to use}

\item{type}{geographies can be listed as various types. Here we are just
using Census geographies and Administrative geographies. This affects the
beginning of the query URL. "census" is the default.}

\item{within_level}{The area level variable name associated with the
locations filter. e.g. \code{"cty19nm"} or \code{"rgn19nm"}}

\item{within}{A place name, or list of place names or codes, to filter the
data by. If nothing is stipulated then the full unfiltered table will be
returned}

\item{fields}{The fields of the data to be returned. Defaults to \code{"*"}
(all); can instead be a set of column names/variables.}

\item{return_geometry}{Whether to return a geometry}

\item{sr}{The (EPSG) spatial reference of any returned geometry.
4326 ("WGS 84") by default. Can be specified as numeric or character.}
}
\value{
a string that should function as a valid API query
}
\description{
A function called by create_custom_lookup.R and geo_get.R to build a
valid query
}
\examples{
build_api_query(ref = 4,
  within_level = "cauth20nm",
  within = "Greater Manchester",
  fields = c("lad20cd", "lad20nm", "cauth20cd", "cauth20nm")
)
build_api_query(ref = 12,
  within_level = "lad20nm",
  within = c(
    "Cheltenham", "Gloucester",
    "Stroud", "Cotswold",
    "Tewkesbury", "Forest of Dean"
  ),
  fields = c("lad20cd", "lad20nm")
)
}
