% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_api_query.R
\name{build_api_query}
\alias{build_api_query}
\title{Helper Function to Build Queries to the Open Geography API}
\usage{
build_api_query(
  table_code_ref,
  type = "census",
  server = "feature",
  search_within,
  locations = NULL,
  fields = "*"
)
}
\arguments{
\item{table_code_ref}{an integer, passed by the calling function, that indicates which table_code to use}

\item{type}{geographies can be listed as various types. Here we are just using Census geographies and Administrative geographies. This affects the beginning of the query URL. Census is the default, but Admin can be passed instead where necessary.}

\item{server}{Some API queries (lookup tables) require the Feature server, others (boundaries) require the Map server}

\item{search_within}{The area level variable name associated with the locations filter. e.g. \code{"cty19nm"} or \code{"rgn19nm"}}

\item{locations}{A place name, or list of place names, to filter the data by. If nothing is stipulated then the full unfiltered table will be returned}

\item{fields}{The fields of the data to be returned. Defaults to \code{"*"} (all); can instead be a set of column names/variables.}
}
\value{
a string that should function as a valid API query
}
\description{
A function called by create_custom_lookup.R and geo_get.R to build a valid query
}
\examples{
build_api_query(
  table_code = "Wards_December_2019_Boundaries_EW_BFC",
  type = "admin",
  server = "map"
)
}
