% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_custom_lookup.R
\name{create_custom_lookup}
\alias{create_custom_lookup}
\title{Title}
\usage{
create_custom_lookup(
  bounds_level,
  within,
  within_level,
  include_msoa = NULL,
  return_style = "tidy"
)
}
\arguments{
\item{return_style}{}
}
\value{

}
\description{
Title
}
