% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_get.R
\name{geo_get}
\alias{geo_get}
\title{Get boundaries for small areas within a local authority area}
\usage{
geo_get(
  bounds_level,
  within,
  within_level = NULL,
  within_cd = FALSE,
  include_msoa = NULL,
  return_style = "tidy",
  include_welsh_names = NULL,
  return_boundaries = TRUE,
  return_centroids = FALSE,
  centroid_fields = FALSE,
  shape_fields = FALSE,
  spatial_ref = 4326,
  quiet_read = TRUE
)
}
\arguments{
\item{bounds_level}{The lowest level at which to return codes and names, eg
"LSOA". Has to be one of "lsoa", "msoa", "wd/ward", "lad",
"cty/county". Case-insensitive.}

\item{within}{The name of a geographic area to filter by eg "Swindon",
"Gloucestershire", "Wales".}

\item{within_level}{Upper geographic level to filter at. eg if filtering to
find all LSOAs in a local authority, \code{within_level} will be "lad". Has
to be one of "wd/ward", "lad", "cty/county", "utla/upper", "rgn/region",
"cauth" or "ctry/country". Case-insensitive. Not all combinations of
\code{bounds_level} and \code{within_level} make sense or are possible! NB
"county" includes metropolitan counties such as "Inner London", "Tyne and
Wear" and "West Midlands".}

\item{within_cd}{Usually you'll build the query with a place name to search
within. But sometimes you may wish to pass in a vector of area codes
instead (if that's all you have, or more likely if you are querying within
wards, which don't have unique names (there's a lot of Abbey wards in
England!)). If you're passing in area codes not names, set this to TRUE.}

\item{include_msoa}{If \code{bounds_level} = LSOA and return_style is "tidy",
whether to also include MSOA columns (in "tidy" return style). If
\code{bounds_level} is MSOA, this will be forced to \code{TRUE}.}

\item{return_style}{"tidy" (the default) means all available columns between
\code{bounds_level} and \code{within_level} will be returned, but with any
empty columns removed. "simple" means that only the code and name (cd and
nm) columns for \code{bounds_level} and \code{within_level} are returned -
other columns are omitted. "minimal" means 'only return the columns for
\code{bounds_level}'.}

\item{include_welsh_names}{Only makes a difference when \code{bounds_level} =
msoa, or when \code{bounds_level} = lsoa and \code{return_style} = "tidy".
\code{FALSE} returns no Welsh language columns. \code{TRUE} attempts to
return Welsh language LAD and MSOA names where relevant. \code{NULL} (the
default) means that an educated decision will be made by the program,
based on whether any of the areas returned have "^W" codes.}

\item{return_boundaries}{whether to retrieve object boundaries data from
the API. Default \code{TRUE}. If \code{return_boundaries} and
\code{return_centroids} are both \code{FALSE}, a plain summary df
without geometry will be returned.}

\item{return_centroids}{whether to retrieve area centroids instead of
boundaries. Default \code{FALSE}. If set to TRUE then it will override
\code{return_boundaries} whether that was set TRUE or otherwise. If
\code{return_boundaries} and \code{return_centroids} are both \code{FALSE},
a plain summary data frame without geometry will be returned.}

\item{centroid_fields}{Boolean, default FALSE. Whether to include BNG
eastings, northings, lat and long fields in the return when returning
boundaries.}

\item{shape_fields}{Boolean, default FALSE. Whether to include
Shape__Area and Shape__Length fields in the return when returning
boundaries.}

\item{spatial_ref}{The (EPSG) spatial reference of any returned geometry.
Default value: 4326 ("WGS 84"). This parameter is ignored peacefully if
no geometry is returned/returnable, eg lookup queries}

\item{quiet_read}{Controls quiet parameter to sf::st_read}
}
\value{
a data frame or an sf (simple features) object (data frame
with geometries)
}
\description{
The ONS OpenGeography Portal (\url{https://geoportal.statistics.gov.uk/})
is a great resource for area boundary and data lookups within the UK.
This program focuses on areas within England and Wales only, initially.
}
\details{
\emph{I want a better name for this function! Suggestions welcome...}
}
\examples{
geo_get("wd", "Swindon", "lad")
geo_get("msoa", "Swansea", "lad", centroid_fields = TRUE) \%>\%
  head(10)
geo_get("lsoa", "Zetland", "ward", shape_fields = TRUE)
geo_get(bounds_level = "lad",
  within = "Gloucestershire",
  within_level = "cty",
  return_style = "simple",
  return_boundaries = FALSE)
}
