% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_get.R
\name{geo_get}
\alias{geo_get}
\title{Get boundaries for small areas within a local authority area}
\usage{
geo_get(
  place = NA,
  search_within = "lad19",
  return_areas = "wd19",
  boundaries = TRUE,
  minimal = FALSE
)
}
\arguments{
\item{place}{Name of an area such as a local authority}

\item{search_within}{Area level of 'place', LAD by default (lad19nm)}

\item{return_areas}{Sub-areas to be retrieved}

\item{boundaries}{Boolean, default TRUE. Whether to return geometry as well as lookup df}

\item{minimal}{Boolean, default FALSE. Whether to return all columns of lookup df or just the columns for the specified area and sub-area levels}
}
\value{
an sf (simple features) object
}
\description{
The ONS OpenGeography Portal (https://geoportal.statistics.gov.uk/)
is a great resource for area boundary and data lookups within the UK.
This program focuses on areas within England and Wales only, initially.
}
\examples{

}
