% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundr-package.R
\name{gregg}
\alias{gregg}
\title{Conveniently wrap a regular expr in glue::glue_data() and pass to \code{grepl()}}
\usage{
gregg(x, rx, ...)
}
\arguments{
\item{x}{A character vector to check}

\item{rx}{A string that after processing by glue_data() will be used as a regex pattern in \code{grepl()}}

\item{...}{Arguments passed onto \code{grepl()}}
}
\description{
Conveniently wrap a regular expr in glue::glue_data() and pass to \code{grepl()}
}
\keyword{internal}
