% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundr-package.R
\name{gregg}
\alias{gregg}
\title{Conveniently wrap a regex in glue::glue and pass it to grepl}
\usage{
gregg(x, rx, ...)
}
\arguments{
\item{x}{A character vector that you want to check}

\item{rx}{A string that when processed by glue() will be used as a regex}

\item{...}{arguments passed onto grepl()}
}
\description{
Conveniently wrap a regex in glue::glue and pass it to grepl
}
\keyword{internal}
