% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_query_info.R
\name{process_query_info}
\alias{process_query_info}
\title{Just another piece of the pipeline}
\usage{
process_query_info(
  query_info,
  within_names,
  within_codes,
  return_width,
  query_opt
)
}
\description{
Just another piece of the pipeline
}
\keyword{internal}
